requirements.txt
polib
python-dotenv
langcodes
google-generativeai
